services/auth_service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyJWT==2.3.0
requests==2.26.0
